ml_backend/Ai chatbot/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.5.2
python-dotenv==1.0.0
langchain==0.1.12
langchain-groq==0.1.5
langsmith==0.1.17
scikit-learn==1.4.0
numpy==1.26.3
faiss-cpu==1.7.4
pypdf==4.0.1